requirements.txt
deezer-py=1.3.7
flask=2.2.2